# backend/docs/swagger.yaml
basePath: /api/v2
definitions:
...
  github_com_norman6464_FreStyle_backend_internal_domain.CompanyApplication:
    properties:
      applicantName:
        type: string
      companyName:
        type: string
      createdAt:
        type: string
      email:
        type: string
      id:
        type: integer
      message:
        type: string
      status:
        type: string
      updatedAt:
        type: string
    type: object
...
  github_com_norman6464_FreStyle_backend_internal_domain.NoteImageUploadURL:
    properties:
      expiresIn:
        type: integer
      key:
        type: string
      url:
        type: string
    type: object
  github_com_norman6464_FreStyle_backend_internal_domain.Notification:
    properties:
      body:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      isRead:
        type: boolean
      title:
        type: string
      type:
        type: string
      userId:
        type: integer
    type: object
  github_com_norman6464_FreStyle_backend_internal_domain.ProfileImageUploadURL:
    properties:
      expiresIn:
        type: integer
      imageUrl:
        type: string
      key:
        type: string
      uploadUrl:
        type: string
    type: object
...
  internal_handler.createCompanyApplicationReq:
    properties:
      applicantName:
        type: string
      companyName:
        type: string
      email:
        type: string
      message:
        type: string
    required:
    - applicantName
    - companyName
    - email
    type: object
...
  internal_handler.errorResponse:
    properties:
      error:
        example: unauthorized
        type: string
    type: object
...
  internal_handler.issueProfileImageReq:
    properties:
      contentType:
        type: string
      fileName:
        type: string
    type: object
  internal_handler.issueUploadURLReq:
    properties:
      contentType:
        type: string
    type: object
...
  internal_handler.updateCompanyApplicationStatusReq:
    properties:
      status:
        type: string
    required:
    - status
    type: object
...
info:
  contact:
    name: FreStyle Engineering
    url: https://normanblog.com
  description: |-
    新卒 IT エンジニア 向け 統合 研修 プラットフォーム の REST API。
    Clean Architecture (port = usecase/repository / adapter = adapter/persistence) で 構築。
  termsOfService: https://normanblog.com/terms
  title: FreStyle Backend API
  version: "2.0"
paths:
...
  /admin/company-applications:
    get:
      description: 受け付けた企業申請を新しい順で返す。super_admin 専用。
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_norman6464_FreStyle_backend_internal_domain.CompanyApplication'
            type: array
        "401":
          description: 未認証
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "403":
          description: super_admin 以外
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "500":
          description: DB 失敗
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      security:
      - CookieAuth: []
      summary: 企業申請一覧（super_admin）
      tags:
      - admin
  /admin/company-applications/{id}/status:
    patch:
      consumes:
      - application/json
      description: 申請を approved / rejected / pending に更新する。super_admin 専用。
      parameters:
      - description: 申請 ID
        in: path
        name: id
        required: true
        type: integer
      - description: status
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.updateCompanyApplicationStatusReq'
      produces:
      - application/json
      responses:
        "204":
          description: 成功（本文なし）
        "400":
          description: id / status 不正
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "401":
          description: 未認証
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "403":
          description: super_admin 以外
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "500":
          description: DB 更新失敗
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      security:
      - CookieAuth: []
      summary: 企業申請の status 更新（super_admin）
      tags:
      - admin
...
  /company-applications:
    post:
      consumes:
      - application/json
      description: ログイン前のユーザーが会社名 / 氏名 / メール / 任意メッセージで利用申請を送る。受理時に super_admin へ通知する。
      parameters:
      - description: 申請内容
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.createCompanyApplicationReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_norman6464_FreStyle_backend_internal_domain.CompanyApplication'
        "400":
          description: バリデーションエラー
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "429":
          description: レート制限超過
          headers:
            Retry-After:
              description: '再試行までの秒数 (例: 60)'
              type: string
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "500":
          description: 内部エラー
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      summary: 企業利用申請（公開 / 認証不要）
      tags:
      - company-applications
...
  /notes/image-upload-url:
    post:
      consumes:
      - application/json
      description: current user 用 の S3 PUT 署名 URL を 発行。 userId は body から 受け取らず middleware
        の current user を 使う (IDOR 対策、 Phase 3 で 修正)。
      parameters:
      - description: contentType (任意)
        in: body
        name: body
        schema:
          $ref: '#/definitions/internal_handler.issueUploadURLReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_norman6464_FreStyle_backend_internal_domain.NoteImageUploadURL'
        "400":
          description: 発行 失敗
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "401":
          description: 未 認証
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      security:
      - CookieAuth: []
      summary: ノート 画像 PUT 署名 URL
      tags:
      - notes
  /notifications:
    get:
      description: current user の 通知 を 作成 日 降順 で 返す。
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_norman6464_FreStyle_backend_internal_domain.Notification'
            type: array
        "400":
          description: DB 取得 失敗
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "401":
          description: 未 認証
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      security:
      - CookieAuth: []
      summary: 通知 一覧
      tags:
      - notifications
  /notifications/{id}/read:
    patch:
      description: 指定 通知 を 既読 に する (所有者 検証 込み)。 同 ハンドラ は 旧 クライアント 互換 で PUT も 同じ パス
        で 受け付ける が、 OpenAPI で は PATCH を 標準 と して 1 つ だけ 表現 する。
      parameters:
      - description: 通知 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: 成功 (本文 なし)
        "400":
          description: DB 失敗
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "401":
          description: 未 認証
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      security:
      - CookieAuth: []
      summary: 通知 単一 既読 化
      tags:
      - notifications
  /notifications/read-all:
    patch:
      description: current user の 全 未読 通知 を まとめて 既読 に する。 同 パス で PUT も 受け付ける 旧 互換
        ルート あり、 OpenAPI で は PATCH のみ 表現。
      produces:
      - application/json
      responses:
        "204":
          description: 成功 (本文 なし)
        "400":
          description: DB 失敗
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "401":
          description: 未 認証
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      security:
      - CookieAuth: []
      summary: 通知 一括 既読 化
      tags:
      - notifications
  /notifications/unread-count:
    get:
      description: current user の 未読 通知 数 を 整数 で 返す (バッジ 表示 用)。
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: DB 失敗
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "401":
          description: 未 認証
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      security:
      - CookieAuth: []
      summary: 未読 通知 数
      tags:
      - notifications
...
  /profile/{userId}/image/presigned-url:
    post:
      consumes:
      - application/json
      description: プロフィール アイコン 用 の S3 PUT 署名 URL を 発行。 "me" / 数字 一致 のみ 許可 (IDOR 対策)。
        body は 任意。
      parameters:
      - description: 数字 ID または 'me'
        in: path
        name: userId
        required: true
        type: string
      - description: fileName / contentType (任意)
        in: body
        name: body
        schema:
          $ref: '#/definitions/internal_handler.issueProfileImageReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_norman6464_FreStyle_backend_internal_domain.ProfileImageUploadURL'
        "400":
          description: 発行 失敗
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "401":
          description: 未 認証
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "403":
          description: 他 user 指定
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      security:
      - CookieAuth: []
      summary: プロフィール 画像 PUT 署名 URL
      tags:
      - profile
...
securityDefinitions:
  CookieAuth:
    description: Cognito 由来 の JWT を HttpOnly Cookie で 送る。 ログイン 後 自動 付与。
    in: header
    name: Cookie
    type: apiKey
swagger: "2.0"